requirements.txt
coverage==7.4.3
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ipython==8.22.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
stripe==8.6.0
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
requests~=2.31.0
django-cors-headers==4.3.1
flake8==7.0.0